repo restrucutre pt 1
move bot to /bot/
