+++ bot/Pipfile
@@ -1,0 +1,18 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+pylama = "*"
+rope = "*"
+
+[packages]
+twitchio = "*"
+pygame = "*"
+python-socketio = "*"
+phue = "*"
+rgbxy = "*"
+
+[requires]
+python_version = "3.6"
