+++ bot/README.md
